refactor: use unary interceptor
