+++ sdk/executable.go
@@ -15,7 +15,6 @@ import (
 	"github.com/hiero-ledger/hiero-sdk-go/v2/proto/services"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/codes"
-	"google.golang.org/grpc/metadata"
 	"google.golang.org/grpc/status"
 )
 
@@ -247,8 +246,7 @@ func _Execute(client *Client, e Executable) (interface{}, error) {
 
 		var resp interface{}
 
-		md := metadata.Pairs("x-user-agent", client.userAgent)
-		ctx := metadata.NewOutgoingContext(context.Background(), md)
+		ctx := context.Background()
 		var cancel context.CancelFunc
 
 		if e.GetGrpcDeadline() != nil {
